microservices/security-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations
pydantic>=2.5.0
pandas>=2.1.3
python-multipart>=0.0.6
numpy>=1.26.2

# Database dependencies (async)
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0
